chore:test busStopsRepository impl, adds tests jsons, domain model use custom typeAlias for bus stop number

+++ app/src/main/java/com/abrahamcardenes/wawaamarillalimon/data/mappers/BusStops.kt
@@ -7,7 +7,7 @@ fun List<BusStopDto>.toDomain(): List<BusStop> {
     return this.map {
         BusStop(
             addressName = it.addressName,
-            stopNumber = it.stopNumber
+            stopNumber = it.stopNumber.toInt()
         )
     }
 }
